Fixed the issues on Firefox with the navigator of the Lydzje section
Fonts of the navigator in the Lydzje section were bigger when they were
hovered on Firefox. They are supossed to get more weight (bold) when
hovering them, but Firefox makes bold fonts slightly bigger than normal
ones and this made the rest of the links move.

Now, I've just given to those links a bigger width as a default
attribute so they do not move when fonts are bold on Firefox.