Fixed weird behaviour on Firefox when moving sections
Sections were not moved smoothly on Firefox, that was because of the
z-index transition time. It was set to 0s beucase it makes no sense to
set a transition to the z-index attribute, and that worked for all web
browsers I tested but Firefox.

I've just changed the 0s to 0.001s and that worked.